test-data/pyproject.toml
# test-data/pyproject.toml
# Script-based tool, not an installable package

[project]
name = "test-data-generator"
version = "0.1.0"
description = "Test label image generator"
requires-python = ">=3.11"
  dependencies = [
    "pyyaml>=6.0.1",
    "jinja2>=3.1.4",
    "pydantic>=2.9.2",
    "playwright>=1.48.0",
    "ruff>=0.6.0",
  ]

 [project.optional-dependencies]
  dev = [
    "pytest>=8.3.3",
    "pytest-playwright>=0.5.2",
    "ruff>=0.6.0",
  ]

 # This is a virtual/non-package project (just scripts + dependencies)
 # No build system needed - uv will only manage dependencies
 [tool.uv]
 package = false

 [tool.ruff]
 line-length = 88
 target-version = "py311"

 [tool.ruff.lint]
  select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
  ]
  ignore = [
    "E501", # line too long, handled by black
  ]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
